move table reader and functions to common
