+++ src/connector/src/connector_common/postgres.rs
@@ -12,16 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::HashMap;
 use std::fmt;
 
 use anyhow::anyhow;
 use openssl::ssl::{SslConnector, SslMethod, SslVerifyMode};
 use postgres_openssl::MakeTlsConnector;
+use risingwave_common::catalog::{ColumnDesc, ColumnId};
+use risingwave_common::types::{DataType, ScalarImpl, StructType};
+use sea_schema::postgres::def::{ColumnType, TableInfo};
+use sea_schema::postgres::discovery::SchemaDiscovery;
 use serde_derive::Deserialize;
+use sqlx::postgres::{PgConnectOptions, PgSslMode};
+use sqlx::PgPool;
 use thiserror_ext::AsReport;
 use tokio_postgres::{Client as PgClient, NoTls};
 
 use super::maybe_tls_connector::MaybeMakeTlsConnector;
+use crate::error::ConnectorResult;
+use crate::source::cdc::external::ExternalTableConfig;
 
 #[derive(Debug, Clone, PartialEq, Deserialize)]
 #[serde(rename_all = "lowercase")]
@@ -45,6 +54,102 @@ pub enum SslMode {
 impl Default for SslMode {
     fn default() -> Self {
         Self::Preferred
+    }
+}
+
+pub struct PostgresExternalTable {
+    column_descs: Vec<ColumnDesc>,
+    pk_names: Vec<String>,
+}
+
+impl PostgresExternalTable {
+    pub async fn connect(config: ExternalTableConfig) -> ConnectorResult<Self> {
+        tracing::debug!("connect to postgres external table");
+        let mut options = PgConnectOptions::new()
+            .username(&config.username)
+            .password(&config.password)
+            .host(&config.host)
+            .port(config.port.parse::<u16>().unwrap())
+            .database(&config.database)
+            .ssl_mode(match config.ssl_mode {
+                SslMode::Disabled => PgSslMode::Disable,
+                SslMode::Preferred => PgSslMode::Prefer,
+                SslMode::Required => PgSslMode::Require,
+                SslMode::VerifyCa => PgSslMode::VerifyCa,
+                SslMode::VerifyFull => PgSslMode::VerifyFull,
+            });
+
+        if config.ssl_mode == SslMode::VerifyCa || config.ssl_mode == SslMode::VerifyFull {
+            if let Some(ref root_cert) = config.ssl_root_cert {
+                options = options.ssl_root_cert(root_cert.as_str());
+            }
+        }
+
+        let connection = PgPool::connect_with(options).await?;
+        let schema_discovery = SchemaDiscovery::new(connection, config.schema.as_str());
+        // fetch column schema and primary key
+        let empty_map = HashMap::new();
+        let table_schema = schema_discovery
+            .discover_table(
+                TableInfo {
+                    name: config.table.clone(),
+                    of_type: None,
+                },
+                &empty_map,
+            )
+            .await?;
+
+        let mut column_descs = vec![];
+        for col in &table_schema.columns {
+            let data_type = type_to_rw_type(&col.col_type)?;
+            let column_desc = if let Some(ref default_expr) = col.default {
+                // parse the value of "column_default" field in information_schema.columns,
+                // non number data type will be stored as "'value'::type"
+                let val_text = default_expr
+                    .0
+                    .split("::")
+                    .map(|s| s.trim_matches('\''))
+                    .next()
+                    .expect("default value expression");
+
+                match ScalarImpl::from_text(val_text, &data_type) {
+                    Ok(scalar) => ColumnDesc::named_with_default_value(
+                        col.name.clone(),
+                        ColumnId::placeholder(),
+                        data_type.clone(),
+                        Some(scalar),
+                    ),
+                    Err(err) => {
+                        tracing::warn!(error=%err.as_report(), "failed to parse postgres default value expression, only constant is supported");
+                        ColumnDesc::named(col.name.clone(), ColumnId::placeholder(), data_type)
+                    }
+                }
+            } else {
+                ColumnDesc::named(col.name.clone(), ColumnId::placeholder(), data_type)
+            };
+            column_descs.push(column_desc);
+        }
+
+        if table_schema.primary_key_constraints.is_empty() {
+            return Err(anyhow!("Postgres table doesn't define the primary key").into());
+        }
+        let mut pk_names = vec![];
+        table_schema.primary_key_constraints.iter().for_each(|pk| {
+            pk_names.extend(pk.columns.clone());
+        });
+
+        Ok(Self {
+            column_descs,
+            pk_names,
+        })
+    }
+
+    pub fn column_descs(&self) -> &Vec<ColumnDesc> {
+        &self.column_descs
+    }
+
+    pub fn pk_names(&self) -> &Vec<String> {
+        &self.pk_names
     }
 }
 
@@ -141,4 +246,91 @@ pub async fn create_pg_client(
     });
 
     Ok(client)
+}
+
+pub fn type_to_rw_type(col_type: &ColumnType) -> ConnectorResult<DataType> {
+    let dtype = match col_type {
+        ColumnType::SmallInt | ColumnType::SmallSerial => DataType::Int16,
+        ColumnType::Integer | ColumnType::Serial => DataType::Int32,
+        ColumnType::BigInt | ColumnType::BigSerial => DataType::Int64,
+        ColumnType::Money | ColumnType::Decimal(_) | ColumnType::Numeric(_) => DataType::Decimal,
+        ColumnType::Real => DataType::Float32,
+        ColumnType::DoublePrecision => DataType::Float64,
+        ColumnType::Varchar(_) | ColumnType::Char(_) | ColumnType::Text => DataType::Varchar,
+        ColumnType::Bytea => DataType::Bytea,
+        ColumnType::Timestamp(_) => DataType::Timestamp,
+        ColumnType::TimestampWithTimeZone(_) => DataType::Timestamptz,
+        ColumnType::Date => DataType::Date,
+        ColumnType::Time(_) | ColumnType::TimeWithTimeZone(_) => DataType::Time,
+        ColumnType::Interval(_) => DataType::Interval,
+        ColumnType::Boolean => DataType::Boolean,
+        ColumnType::Point => DataType::Struct(StructType::new(vec![
+            ("x", DataType::Float32),
+            ("y", DataType::Float32),
+        ])),
+        ColumnType::Uuid => DataType::Varchar,
+        ColumnType::Xml => DataType::Varchar,
+        ColumnType::Json => DataType::Jsonb,
+        ColumnType::JsonBinary => DataType::Jsonb,
+        ColumnType::Array(def) => {
+            let item_type = match def.col_type.as_ref() {
+                Some(ty) => type_to_rw_type(ty.as_ref())?,
+                None => {
+                    return Err(anyhow!("ARRAY type missing element type").into());
+                }
+            };
+
+            DataType::List(Box::new(item_type))
+        }
+        ColumnType::PgLsn => DataType::Int64,
+        ColumnType::Cidr
+        | ColumnType::Inet
+        | ColumnType::MacAddr
+        | ColumnType::MacAddr8
+        | ColumnType::Int4Range
+        | ColumnType::Int8Range
+        | ColumnType::NumRange
+        | ColumnType::TsRange
+        | ColumnType::TsTzRange
+        | ColumnType::DateRange
+        | ColumnType::Enum(_) => DataType::Varchar,
+
+        ColumnType::Line => {
+            return Err(anyhow!("LINE type not supported").into());
+        }
+        ColumnType::Lseg => {
+            return Err(anyhow!("LSEG type not supported").into());
+        }
+        ColumnType::Box => {
+            return Err(anyhow!("BOX type not supported").into());
+        }
+        ColumnType::Path => {
+            return Err(anyhow!("PATH type not supported").into());
+        }
+        ColumnType::Polygon => {
+            return Err(anyhow!("POLYGON type not supported").into());
+        }
+        ColumnType::Circle => {
+            return Err(anyhow!("CIRCLE type not supported").into());
+        }
+        ColumnType::Bit(_) => {
+            return Err(anyhow!("BIT type not supported").into());
+        }
+        ColumnType::VarBit(_) => {
+            return Err(anyhow!("VARBIT type not supported").into());
+        }
+        ColumnType::TsVector => {
+            return Err(anyhow!("TSVECTOR type not supported").into());
+        }
+        ColumnType::TsQuery => {
+            return Err(anyhow!("TSQUERY type not supported").into());
+        }
+        ColumnType::Unknown(name) => {
+            // NOTES: user-defined enum type is classified as `Unknown`
+            tracing::warn!("Unknown Postgres data type: {name}, map to varchar");
+            DataType::Varchar
+        }
+    };
+
+    Ok(dtype)
 }
